This is a contribution from Ribose Inc (@riboseinc).

    Algorithm uses 4 tables of precalculated CRC24 values, thanks to which
    it can process in parallel 32 bits of data. This tric doubles performance
    Further improvements are possible.

    Results - (tested with RNP) processing 1GB armor data

    ```
    OLD: rnp --enarmor=msg /tmp/1gb.rnd --output  4.48s user 0.89s system 98% cpu 5.429 total
    NEW: rnp --enarmor=msg /tmp/1gb.rnd --output  2.38s user 0.86s system 79% cpu 4.089 total

    OLD: rnp --dearmor out.xxx --output out.d  5.58s user 0.65s system 98% cpu 6.338 total
    NEW: rnp --dearmor out.xxx --output out.d  3.28s user 0.84s system 96% cpu 4.275 total
    ```

Please let me know if something needs to be changed in the review.