@randombit 2017-10-28
Nice, thanks. Can you add a copyright statement at the top of the file? The tests are crashing in CI, presumably the overwrite I commented on. Also if you could add a brief comment at the top explaining the theory of operation that would be nice. Just 1 or 2 lines is fine. Basically T1[i] is (2^8 * T0[i]) in GF(2^24) modulo the CRC24 polynomial, right?