@flowher 2017-10-28
Okis, adding it now. Yes that's the basic idea.