@flowher 2017-10-29
Errors seem unrelated

```
SM2/SM3 decryption ran 10 tests in 187.50 msec 2 FAILED
Failure 1: SM2/SM3 decryption: incorrect ciphertext different produced matching
Failure 2: SM2/SM3 decryption unexpected result for used corrupted ciphertext
```