@ronaldtse 2017-10-29
@flowher @randombit the SM2/SM3 issue may be due to the changed ASN.1 output of SM2?