@flowher 2017-10-29
Please don't merge it even if it's OK. Final merge must be done from riboseinc account. It's not possible to change source branch, so after everything is OK I'll open new PR with same code from ribose branch