@randombit 2017-10-29
Looks good to me. SM2 problem is unrelated. Likely we are flipping a bit in some region that gets ignored during decryption. I'll check it out.