@flowher 2017-10-29
Okis, then I'm closing this PR and will open exact same from ribose